Added User Friendly-derps
Added prompt and print output for usability.

+++ rpi-webdash-setup.sh
@@ -1,4 +1,4 @@
-# prior to running this script, please perform the following actions
+# --- pre-script actions ---
 
 # > extend root file system vol
 # sudo raspi-config 
@@ -31,25 +31,42 @@
 # hdmi_mode=39
 
 # -- BEGIN SCRIPT
+clear
+echo rpi-webdash setup script
+echo created 2-06-2015
+echo **NOTICE**: 
+echo Please make sure you have read the contents of 
+echo this script before running for pre and post 
+echo config instructions.. (sudo nano rpi-webdash-setup.sh)
+echo ---
+read -p "Are you sure you want to continue? <y/n> " prompt
+if [[ $prompt == "y" || $prompt == "Y" || $prompt == "yes" || $prompt == "Yes" ]]
+then
+	# setup NTP Time
+	echo Install and setup NTP Server 
+	sudo apt-get install ntpdate
+	sudo ntpdate -u ntp.ubuntu.com
 
-# setup NTP Time
-sudo apt-get install ntpdate
-sudo ntpdate -u ntp.ubuntu.com
+	# update raspbian and perform firmware update
+	echo Update raspbian and perform firmware upgrade
+	sudo apt-get update
+	sudo apt-get upgrade
+	sudo apt-get dist-upgrade
 
-# update raspbian and perform firmware update
-sudo apt-get update
-sudo apt-get upgrade
-sudo apt-get dist-upgrade
+	# install git
+	echo Setup git-core
+	sudo apt-get install git-core
 
-# install git
-sudo apt-get install git-core
-
-# install chromium with mscorefonts for readability
-sudo apt-get install chromium ttf-mscorefonts-installer
+	# install chromium with mscorefonts for readability
+	echo Install Chromium with mscorefonts
+	sudo apt-get install chromium ttf-mscorefonts-installer
+else
+	exit 0
+fi
 
 # -- END
 
-# post script actions
+# --- post-script actions ---
 
 # > autostart chromium, load webpage in kiosk and incognito mode to remove recovery tab on power interrupt
 # sudo nano /etc/xdg/lxsession/LXDE-pi/autostart
